fix(ut): unified tests based on original file name
