+++ test/unit/src/utils/general.test.ts
@@ -1,4 +1,4 @@
-import { compareObjectShape } from '../../../../src/utils'
+import { compareObjectShape, generateArraySchema, generateObjectSchema } from '../../../../src/utils'
 
 test('compareObjectShape() > two identical object should return true', () => {
   const idol = {
@@ -252,6 +252,7 @@ test('compareObjectShape() > should return error object on array type mismatch',
   expect(isValid).toBe(false)
   expect(JSON.stringify(expectedError) === JSON.stringify(error)).toBe(true)
 })
+
 test('compareObjectShape() > should not fail when property possess falsy value', () => {
   const idol = {
     timestamp: new Date(),
@@ -288,4 +289,75 @@ test('compareObjectShape() > should not fail when property possess falsy value',
 
   expect(isValid).toBe(true)
   expect(error).toBe(undefined)
+})
+
+test('generateArraySchema() > should generate proper array schema', () => {
+  const num = [1, 3, 2]
+  const str = ['john', 'doe']
+  const fn = [(str) => str]
+  const instanceObj = [new Date(), new Date()]
+  const literalObj = [{}, {}]
+  const any = [1, false, 'doe']
+  const dimensional = [
+    [1, 2, 3],
+    [1, 1],
+  ]
+
+  expect(generateArraySchema(num)).toBe('number[]')
+  expect(generateArraySchema(str)).toBe('string[]')
+  expect(generateArraySchema(fn)).toBe('function[]')
+  expect(generateArraySchema(instanceObj)).toBe('object[]')
+  expect(generateArraySchema(literalObj)).toBe('{}[]')
+  expect(generateArraySchema(any, { diversity: true })).toBe('any[]')
+
+  try {
+    generateArraySchema(any)
+  } catch (e) {
+    expect(e.message).toBe(
+      'Array schema generation does not allowed type diversities in an array unless specified'
+    )
+  }
+  expect(generateArraySchema(dimensional)).toBe('array[]')
+})
+
+test('generateObjectSchema() > should generate proper object schema', () => {
+  const obj = {
+    timestamp: new Date(),
+    fn: (str) => str,
+    falsy: -0,
+    id: 1,
+    dummyObj: { id: 1, name: 'john doe', NRC: '23403492340' }, // this information is purely made up
+    nestedObj: {
+      numArr: [1, 3, 4],
+      strArr: ['name1', 'name2'],
+      multiDiArr: [[12, 32]],
+      nestedArr: [{ id: 1 }, { id: 2 }],
+      diverseArr: [1, false, 'str'],
+      id: 1,
+      debug: true,
+    },
+  }
+
+  const expectedSchema = {
+    timestamp: 'object',
+    fn: 'function',
+    falsy: 'number',
+    id: 'number',
+    dummyObj: { id: 'number', name: 'string', NRC: 'string' },
+    nestedObj: {
+      numArr: 'number[]',
+      strArr: 'string[]',
+      multiDiArr: 'array[]',
+      nestedArr: '{}[]',
+      diverseArr: 'any[]',
+      id: 'number',
+      debug: 'boolean',
+    },
+  }
+
+  const generatedSchema = generateObjectSchema(obj, { arrTypeDiversity: true })
+
+  const isEqual = JSON.stringify(generatedSchema) === JSON.stringify(expectedSchema)
+
+  expect(isEqual).toBe(true)
 })
